Added JobService, StockService and replaced Double on BigDecimal

+++ src/main/kotlin/com/app/stocks_service/repository/entity/CompanyDeltaEntity.kt
@@ -1,6 +1,8 @@
 package com.app.stocks_service.repository.entity
 
+import java.math.BigDecimal
+
 data class CompanyDeltaEntity(
     val name: String,
-    val delta: Double
+    val delta: BigDecimal
 )
