+++ src/main/kotlin/com/app/stocks_service/repository/entity/CompanyVolumeEntity.kt
@@ -1,6 +1,8 @@
 package com.app.stocks_service.repository.entity
 
+import java.math.BigDecimal
+
 data class CompanyVolumeEntity(
     val name: String,
-    val volume: Double
+    val volume: BigDecimal
 )
